Cleanup and visualizer
